requirements.txt
flask
flask-cors
torch
transformers
gTTS
deep-translator
pillow
